Create delete.php
This PHP script connects to a MySQL database and deletes a record from the "To do list"  table based on the ID provided in the request. If the deletion is successful, it redirects the user to the "index2.php" page. However, it's important to note that the code is vulnerable to SQL injection and should be modified to use prepared statements for better security.